# src/swagger-doc.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/prices": {
      "post": {
        "operationId": "PricesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "The data to create new price",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePriceDto"
              },
              "examples": {
                "single": {
                  "value": {
                    "price": 300,
                    "top": 500,
                    "bottom": 200,
                    "currency": "EUR",
                    "cityId": 1,
                    "productId": 1,
                    "yearId": 1,
                    "priceType": "HISTORICAL"
                  }
                },
                "multiple": {
                  "value": [
                    {
                      "price": 300,
                      "top": 500,
                      "bottom": 200,
                      "currency": "EUR",
                      "cityId": 1,
                      "productId": 1,
                      "yearId": 1,
                      "priceType": "HISTORICAL"
                    },
                    {
                      "price": 200,
                      "top": 500,
                      "bottom": 200,
                      "currency": "EUR",
                      "cityId": 1,
                      "productId": 1,
                      "yearId": 1,
                      "priceType": "HISTORICAL"
                    }
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created a price",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceDto"
                },
                "examples": {
                  "single": {
                    "summary": "Price DTO",
                    "value": {
                      "id": 1,
                      "price": 300,
                      "top": 500,
                      "bottom": 200,
                      "currency": "EUR",
                      "cityId": 1,
                      "productId": 1,
                      "yearId": 1,
                      "priceType": "HISTORICAL",
                      "createdAt": "2025-03-26T19:50:30.809Z",
                      "updatedAt": "2025-03-26T19:50:30.809Z"
                    }
                  },
                  "multiple": {
                    "summary": "Count",
                    "value": {
                      "count": 100
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Create the new price.",
        "tags": [
          "prices"
        ]
      },
      "get": {
        "operationId": "PricesController_getAll",
        "parameters": [
          {
            "name": "currency",
            "required": false,
            "in": "query",
            "schema": {
              "example": "EUR",
              "type": "string"
            }
          },
          {
            "name": "cityId",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "productId",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "yearId",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "priceType",
            "required": false,
            "in": "query",
            "schema": {
              "example": "HISTORICAL",
              "type": "string",
              "enum": [
                "CURRENT",
                "HISTORICAL"
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Items to skip",
            "schema": {
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Field to sort by",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "Sort order: asc or desc",
            "schema": {
              "example": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved prices.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PricePaginationDto"
                  }
                },
                "examples": {
                  "application/json": {
                    "summary": "price DTO array",
                    "value": {
                      "data": [
                        {
                          "id": 1,
                          "price": 300,
                          "top": 500,
                          "bottom": 200,
                          "currency": "EUR",
                          "cityId": 1,
                          "productId": 1,
                          "yearId": 1,
                          "priceType": "HISTORICAL",
                          "createdAt": "2025-03-26T19:50:30.809Z",
                          "updatedAt": "2025-03-26T19:50:30.809Z"
                        },
                        {
                          "id": 2,
                          "price": 400,
                          "top": 500,
                          "bottom": 200,
                          "currency": "EUR",
                          "cityId": 1,
                          "productId": 1,
                          "yearId": 1,
                          "priceType": "HISTORICAL",
                          "createdAt": "2025-03-26T19:50:30.809Z",
                          "updatedAt": "2025-03-26T19:50:30.809Z"
                        }
                      ],
                      "total": 100,
                      "limit": 10,
                      "page": 1
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Return all prices.",
        "tags": [
          "prices"
        ]
      },
      "put": {
        "operationId": "PricesController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "The data to update price",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePriceDto"
              },
              "examples": {
                "single": {
                  "value": {
                    "id": 1,
                    "price": 300,
                    "currency": "EUR",
                    "cityId": 1,
                    "productId": 1,
                    "yearId": 1,
                    "priceType": "HISTORICAL"
                  }
                },
                "multiple": {
                  "value": [
                    {
                      "id": 1,
                      "price": 300,
                      "top": 500,
                      "bottom": 200,
                      "currency": "EUR",
                      "cityId": 1,
                      "productId": 1,
                      "yearId": 1,
                      "priceType": "HISTORICAL"
                    },
                    {
                      "id": 2,
                      "price": 300,
                      "top": 500,
                      "bottom": 200,
                      "currency": "EUR",
                      "cityId": 1,
                      "productId": 1,
                      "yearId": 1,
                      "priceType": "HISTORICAL"
                    }
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated prices.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PriceDto"
                  }
                },
                "examples": {
                  "single": {
                    "summary": "Single updated price",
                    "value": {
                      "id": 1,
                      "price": 300,
                      "top": 500,
                      "bottom": 200,
                      "currency": "EUR",
                      "cityId": 1,
                      "productId": 1,
                      "yearId": 1,
                      "priceType": "HISTORICAL",
                      "createdAt": "2025-03-26T19:50:30.809Z",
                      "updatedAt": "2025-03-26T19:50:30.809Z"
                    }
                  },
                  "multiple": {
                    "summary": "Multiple updated prices",
                    "value": [
                      {
                        "id": 1,
                        "price": 300,
                        "top": 500,
                        "bottom": 200,
                        "currency": "EUR",
                        "cityId": 1,
                        "productId": 1,
                        "yearId": 1,
                        "priceType": "HISTORICAL",
                        "createdAt": "2025-03-26T19:50:30.809Z",
                        "updatedAt": "2025-03-26T19:50:30.809Z"
                      },
                      {
                        "id": 2,
                        "price": 400,
                        "top": 500,
                        "bottom": 200,
                        "currency": "EUR",
                        "cityId": 1,
                        "productId": 1,
                        "yearId": 1,
                        "priceType": "HISTORICAL",
                        "createdAt": "2025-03-26T19:50:30.809Z",
                        "updatedAt": "2025-03-26T19:50:30.809Z"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "summary": "Update prices",
        "tags": [
          "prices"
        ]
      }
    },
    "/prices/{id}": {
      "get": {
        "operationId": "PricesController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved price by id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceDto"
                },
                "examples": {
                  "application/json": {
                    "summary": "Price DTO",
                    "value": {
                      "id": 1,
                      "price": 300,
                      "top": 500,
                      "bottom": 200,
                      "currency": "EUR",
                      "cityId": 1,
                      "productId": 1,
                      "yearId": 1,
                      "priceType": "HISTORICAL",
                      "createdAt": "2025-03-26T19:50:30.809Z",
                      "updatedAt": "2025-03-26T19:50:30.809Z"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Return price by id.",
        "tags": [
          "prices"
        ]
      },
      "delete": {
        "operationId": "PricesController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted price by id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceDto"
                },
                "examples": {
                  "application/json": {
                    "summary": "Price DTO",
                    "value": {
                      "id": 1,
                      "price": 300,
                      "top": 500,
                      "bottom": 200,
                      "currency": "EUR",
                      "cityId": 1,
                      "productId": 1,
                      "yearId": 1,
                      "priceType": "HISTORICAL",
                      "createdAt": "2025-03-26T19:50:30.809Z",
                      "updatedAt": "2025-03-26T19:50:30.809Z"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Delete price by id.",
        "tags": [
          "prices"
        ]
      }
    }
  },
  "info": {
    "title": "My API",
    "description": "The API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreatePriceDto": {
        "type": "object",
        "properties": {
          "price": {
            "type": "number",
            "description": "The price value"
          },
          "top": {
            "type": "number",
            "description": "Top of the price range"
          },
          "bottom": {
            "type": "number",
            "description": "Bottom of the price range"
          },
          "currency": {
            "type": "string",
            "description": "The currency of the price"
          },
          "cityId": {
            "type": "number",
            "description": "The id of the city (cities table foreign key)"
          },
          "productId": {
            "type": "number",
            "description": "The id of the product (products table foreign key)"
          },
          "yearId": {
            "type": "number",
            "description": "The id of the year (years table foreign key)"
          },
          "priceType": {
            "type": "string",
            "description": "Price Type",
            "enum": [
              "CURRENT",
              "HISTORICAL"
            ]
          }
        },
        "required": [
          "price",
          "currency",
          "cityId",
          "productId",
          "yearId",
          "priceType"
        ]
      },
      "PriceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "The unique identifier of the price"
          },
          "price": {
            "type": "number",
            "description": "The price value"
          },
          "top": {
            "type": "number",
            "description": "Top of the price range"
          },
          "bottom": {
            "type": "number",
            "description": "Bottom of the price range"
          },
          "currency": {
            "type": "string",
            "description": "The currency of the price"
          },
          "cityId": {
            "type": "number",
            "description": "The id of the city (cities table foreign key)"
          },
          "productId": {
            "type": "number",
            "description": "The id of the product (products table foreign key)"
          },
          "yearId": {
            "type": "number",
            "description": "The id of the year (years table foreign key)"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date when the record was created",
            "example": "2025-03-26T12:00:00Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date when the record was last updated",
            "example": "2025-03-26T12:00:00Z"
          },
          "priceType": {
            "type": "string",
            "description": "Price Type",
            "enum": [
              "CURRENT",
              "HISTORICAL"
            ]
          }
        },
        "required": [
          "id",
          "price",
          "currency",
          "cityId",
          "productId",
          "yearId",
          "priceType"
        ]
      },
      "Object": {
        "type": "object",
        "properties": {}
      },
      "PricePaginationDto": {
        "type": "object",
        "properties": {
          "data": {
            "description": "Array of prices matching the query filters",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PriceDto"
            }
          },
          "total": {
            "type": "number",
            "description": "Total number of prices matching the query filters",
            "example": 100
          },
          "limit": {
            "type": "number",
            "description": "Limit of items per page",
            "example": 10
          },
          "offset": {
            "type": "number",
            "description": "Offset of items to skip for pagination",
            "example": 1
          }
        },
        "required": [
          "data",
          "total",
          "limit",
          "offset"
        ]
      }
    }
  }
}